Fix %jd conversions passed off_t arguments
%jd takes an intmax_t, but every use in the tree passed an off_t
(or an off_t * in the sscanf calls).  The two are distinct types:
where they differ, as on macOS (off_t is long long, intmax_t is
long), clang warns on every such call, and passing the wrong type
to a variadic function is undefined behavior even when the sizes
happen to match.

Cast the printf-style arguments to intmax_t and read the sscanf
targets into intmax_t variables; <stdint.h> joins the axel.h
includes for the type.  Twelve call sites across eight files, no
behavior change on LP64 Linux where both types are long.

With this, a default -Werror git build compiles on macOS
(Darwin 25.5, Apple clang).

+++ src/axel.c
@@ -180,7 +180,7 @@ axel_new(conf_t *conf, int count, const search_t *res)
 			char hsize[32];
 			axel_size_human(hsize, sizeof(hsize), axel->size);
 			axel_message(axel, _("File size: %s (%jd bytes)"),
-				     hsize, axel->size);
+				     hsize, (intmax_t)axel->size);
 		} else {
 			axel_message(axel, _("File size: unavailable"));
 		}
@@ -854,8 +854,8 @@ axel_divide(axel_t *axel)
 #ifndef NDEBUG
 	for (int i = 0; i < axel->conf->num_connections; i++) {
 		printf(_("Downloading %jd-%jd using conn. %i\n"),
-		       axel->conn[i].currentbyte,
-		       axel->conn[i].lastbyte, i);
+		       (intmax_t)axel->conn[i].currentbyte,
+		       (intmax_t)axel->conn[i].lastbyte, i);
 	}
 #endif
 }
